refactor: extract cache feature
